#376 feat: implement game state loading functionality
## Changes
- Created a LoadGame class to handle loading actors from saved game data.
- Integrated loading logic into SceneActorCreator during actor initialization.
- Enhanced error handling for scene and actor loading processes.

## Reason
- To allow players to resume their game from saved states.
- To improve the overall game experience by ensuring actors are correctly restored.
