+++ apps/client/src/app/probable-waffle/game/scenes/GameProbableWaffleScene.ts
@@ -49,6 +49,7 @@ export default class GameProbableWaffleScene extends ProbableWaffleScene {
   }
 
   create() {
+    const c = this.baseGameData.gameInstance.gameState;
     const hud = this.scene.get<HudProbableWaffle>("HudProbableWaffle") as HudProbableWaffle;
     hud.scene.start();
     hud.initializeWithParentScene(this);
